\name{print.imagedata}
\alias{print.imagedata}
\title{Print information on a given imagedata object}
\description{
  This function outputs information on a given imagedata object.
}
\usage{
\method{print}{imagedata}(x, ...)
}
\arguments{
  \item{x}{The image}
  \item{...}{Ignored}
}

\seealso{
  \code{\link{imagedata}}
}
\keyword{misc}
